requirements.txt
botbuilder-integration-aiohttp>=4.14.0
pandas
tabulate